+++ Pipfile
@@ -31,6 +31,6 @@ python_version = "3.9"
 build = "python setup.py install sdist bdist bdist_wheel"
 lint = "pylint --rcfile .pylintrc src/"
 test = "pytest"
-check-types ="mypy src/pympl/"
-validate = "bash -c 'mypy src/pympl/ && python -m unittest && pylint src/'"
-install-types ="mypy --install-types --non-interactive src/pympl/  "
+check-types ="mypy src/mpyl/"
+validate = "bash -c 'mypy src/mpyl/ && python -m unittest && pylint src/'"
+install-types ="mypy --install-types --non-interactive src/mpyl/  "
