+++ setup.py
@@ -22,9 +22,9 @@ def get_install_requirements():
 
 
 setup(
-    name="pympl",
-    version="0.0.16",
-    description="A python rewrite of mpl-modules",
+    name="mpyl",
+    version="0.0.1",
+    description="Modular Pipeline Library",
     long_description=readme,
     long_description_content_type="text/markdown",
     install_requires=get_install_requirements(),
