+++ tests/steps/test_steps.py
@@ -3,7 +3,7 @@ from io import StringIO
 
 from ruamel.yaml import YAML  # type: ignore
 
-from src.pympl.steps.models import Output, Artifact, ArtifactType
+from src.mpyl.steps.models import Output, Artifact, ArtifactType
 from tests import root_test_path
 
 yaml = YAML()
