+++ tests/test_resources/mpl/BUILD.yml
@@ -1,7 +1,7 @@
-!!python/object:pympl.steps.models.Output
+!!python/object:mpyl.steps.models.Output
 message: Built service
-produced_artifact: !!python/object:pympl.steps.models.Artifact
-  artifact_type: !!python/object/apply:pympl.steps.models.ArtifactType
+produced_artifact: !!python/object:mpyl.steps.models.Artifact
+  artifact_type: !!python/object/apply:mpyl.steps.models.ArtifactType
   - 1
   producing_step: Docker Build
   revision: tag
